Prevent calling 'keyup' on wrong directive scope
Binding the 'keyup' event to the window object results in getting the
wrong scope when the event is fired. It seems that, then, it is always
the scope of the first directive's instance. By binding the event to the
element itself the scope is correctly applied when the event is
triggered.